requirements.txt
apache-airflow==2.7.1
dbt-snowflake==1.5.0
kafka-python==2.0.2
minio==7.1.15
snowflake-connector-python==3.2.0
pandas==2.0.3
faker==19.3.0
python-dotenv==1.0.0
requests==2.31.0